+++ src/Workspace.cpp
@@ -521,6 +521,8 @@ void WorkspaceFolder::registerTypes(const std::vector<std::string>& disabledGlob
         {
             // Clear any set diagnostics
             client->publishDiagnostics({uri, std::nullopt, {}});
+            TextDocument textDocument(Uri::file(resolvedFilePath), "luau", 0, *definitionsContents);
+            definitionsSourceModules.emplace(packageName, std::make_pair(textDocument, result.sourceModule));
         }
         else
         {