+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Sync to upstream Luau 0.698
+- Table properties are now prioritised above other autocomplete entries (again, fixing a dormant bug)
 
 ### Fixed